Refactor file layout
