+++ .storybook/preview.js
@@ -1,4 +1,4 @@
-import '../src/App.css';
+import '../src/components/App.css';
 
 export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },
