+++ src/App.js
@@ -1,5 +1,5 @@
 import logo from './logo.svg';
-import './App.css';
+import './components/App.css';
 
 function App() {
   return (
