+++ src/stories/Button.stories.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Button } from './Button';
+import { Button } from '../components/Button';
 
 export default {
   title: 'Control/Button',
